codes/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
Authlib==1.3.0
Automat==22.10.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
contourpy==1.2.0
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
exceptiongroup==1.2.0
filelock==3.13.1
fonttools==4.49.0
grpcio==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
importlib_resources==6.1.2
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
lxml==5.1.0
matplotlib==3.8.3
networkx==3.2.1
numpy==1.26.3
packaging==23.2
parsel==1.8.1
pillow==10.2.0
portalocker==2.8.2
Protego==0.3.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.7.0
pydantic_core==2.18.1
PyDispatcher==2.0.7
pyOpenSSL==24.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
qdrant-client==1.8.2
queuelib==1.6.2
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
scikit-learn==1.4.0
scipy==1.12.0
Scrapy==2.11.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
threadpoolctl==3.2.0
tiktoken==0.6.0
tldextract==5.1.1
Twisted==23.10.0
typing_extensions==4.10.0
urllib3==2.2.1
validators==0.22.0
w3lib==2.1.2
weaviate-client==3.26.2
zipp==3.17.0
zope.interface==6.2
